Removed collision + added wider hits

+++ ext/drbbdup/drbbdup.h
@@ -125,7 +125,7 @@ typedef struct {
 typedef struct {
 
     uint required_size;
-    uint8_t hit_gen_threshold;
+    uint hit_gen_threshold;
     int dup_limit;
 } drbbdup_fp_settings_t;
